feat(frontend): trigger connections extracted to @agenta/settings-ui
Tools and Triggers turned out to be nearly free of app-layer coupling — between
them they reach @/oss for AlertPopup and one URL helper, nothing else. Their data
already comes from @agenta/entities and @agenta/entity-ui, so antd was the only
thing keeping them out of the package.

The connections list moves first. AlertPopup becomes a `confirm` prop, and the
two actions that need it hide themselves when a host does not pass one, rather
than opening nothing.

OSS keeps a four-line binding at the old path so the Triggers page is unchanged.

Gates: lint 24/24, tsc 0 across settings-ui, oss, ee and mobile.
