+++ web/packages/agenta-settings-ui/src/index.ts
@@ -26,3 +26,7 @@ export {SettingToggleRow, type SettingToggleRowProps} from "./access/SettingTogg
 export {UpgradeNotice, type UpgradeNoticeProps} from "./access/UpgradeNotice"
 export {DomainsSection, type DomainsSectionProps} from "./access/DomainsSection"
 export {SsoProvidersSection, type SsoProvidersSectionProps} from "./access/SsoProvidersSection"
+export {
+    default as TriggerConnectionsSection,
+    type TriggerConnectionsSectionProps,
+} from "./triggers/TriggerConnectionsSection"
